Gateway manifests (#24)
This change moves the gateway manifest into gateway/.

The folder will contain two manifests:

-   dev.xml: pointing to master or other dev locations as needed

-   stable.xml: latest release

With this change, repo usage should rely on refs/tags instead of
expecting a branch to exist. The goal behind this is to keep
the repository with a manageable size as we expect other
components to be added in the future.